Progressing in leaps and bounds!
Dumped the send email page, since we’re not doing that anymore
Now checks for correctness when entering glucose values (int between
20-400)
Actually saves values when you enter them (or forget to)
Working on the summary page screen now. Got it to print out the dates
of each measurement. I think I can make it print them in color if
they’re unusual or in grey if they were skipped.